Move .ts files out of folders
